+++ Projects/ServerFS2/ClientFS2/ConfigurationWriter/SingleTable/PDUPTTable.cs
@@ -16,16 +16,16 @@ namespace ClientFS2.ConfigurationWriter
 			Device = devicePDUDirection.PDUGroupDevice.Device;
 
 			BytesDatabase = new BytesDatabase(Device.DottedPresentationNameAndAddress);
-			BytesDatabase.AddByte((byte)Device.AddressOnShleif, "Адрес");
-			BytesDatabase.AddByte((byte)(Device.ShleifNo - 1), "Шлейф");
-			BytesDatabase.AddByte((byte)Device.Parent.IntAddress, "Адрес прибора");
+			BytesDatabase.AddByte(Device.AddressOnShleif, "Адрес");
+			BytesDatabase.AddByte((Device.ShleifNo - 1), "Шлейф");
+			BytesDatabase.AddByte(Device.Parent.IntAddress, "Адрес прибора");
 			var deviceCode = DriversHelper.GetCodeForFriver(Device.Driver.DriverType);
-			BytesDatabase.AddByte((byte)deviceCode, "Тип ИУ");
+			BytesDatabase.AddByte(deviceCode, "Тип ИУ");
 
 			var tableBase = panelDatabase.PanelDatabase2.LocalZonesTableGroup.Tables.FirstOrDefault(x => x.UID == Device.Zone.UID);
 			var localZoneNo = (tableBase as ZoneTable).BinaryZone.LocalNo;
-			BytesDatabase.AddShort((short)localZoneNo, "Номер зоны");
-			BytesDatabase.AddByte((byte)devicePDUDirection.Device.IntAddress, "Направление");
+			BytesDatabase.AddShort(localZoneNo, "Номер зоны");
+			BytesDatabase.AddByte(devicePDUDirection.Device.IntAddress, "Направление");
 
 			foreach (var tableGroup in panelDatabase.PanelDatabase2.DevicesTableGroups)
 			{
